Rename unavailability server routes
