+++ server/api/index.js
@@ -8,7 +8,7 @@ router.use('/setup', require('./setup'));
 router.use('/integrations', require('./integrations'));
 router.use('/students', require('./students'));
 router.use('/terms', require('./terms'));
-router.use('/unavailability', require('./unavailability'));
+router.use('/unavailabilities', require('./unavailabilities'));
 router.use('/blocks', require('./blocks'));
 router.use('/todos', require('./todos'));
